[release] Install Crystal w/ crystal-lang/install-crystal action
... rather than using a Crystal Docker image.

This documentation
https://crystal-lang.org/reference/1.15/guides/ci/gh-actions.html#using-the-official-docker-image
links to this post
https://forum.crystal-lang.org/t/github-action-to-install-crystal-and-shards-unified-ci-across-linux-macos-and-windows/2837
that cites some advantages of using the GitHub Action. Supposedly one
advantage is that the GitHub action is faster?, though I don't know why
that would be the case, and I'm skeptical it's true. (The cited source
is no longer accessible.) But one advantage that I think is real is that
it allows us to us the probably more up-to-date and more "official" and
supported-by-GitHub GitHub runner image as a starting point.

We were already using the `crystal-lang/install-crystal` GitHub Action
to install Crystal in `ci.yml`. This change updates us to do the same in
`release.yml`.

Also, prepare for release as v1.1.0 (so that we can test that the
release GitHub Action still works).

Also, change "Download source" to "Check out code" to better align and
hint at the name of the underlying action.